+++ resources/views/payment_methods/info_payment_methods.blade.php
@@ -25,7 +25,7 @@
                     <div class="col-12 col-md-6 order-md-2 order-first">
                         <nav class="float-start float-lg-end" style="--bs-breadcrumb-divider: url(&#34;data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' width='8' height='8'%3E%3Cpath d='M2.5 0L1 1.5 3.5 4 1 6.5 2.5 8l4-4-4-4z' fill='%236c757d'/%3E%3C/svg%3E&#34;);" aria-label="breadcrumb">
                             <ol class="breadcrumb">
-                              <li class="breadcrumb-item"><a href="{{ route('app_payment_methods', ['id' => $entreprise->id, 'id2' => $functionalUnit->id]) }}">{{ __('dashboard.payment_methods') }}</a></li>
+                              <li class="breadcrumb-item"><a href="{{ route('app_payment_methods', ['group' => 'payment_method', 'id' => $entreprise->id, 'id2' => $functionalUnit->id]) }}">{{ __('dashboard.payment_methods') }}</a></li>
                               <li class="breadcrumb-item active" aria-current="page">{{ __('payment_methods.payment_method_details') }}</li>
                             </ol>
                         </nav>
@@ -221,47 +221,48 @@
                         </div>
 
                         <div class="row">
+                            @if ($permission_assign || Auth::user()->role->name == "admin" || Auth::user()->role->name == "superadmin")
+                                <div class="col-md-6 mb-3">
+                                    @if ($paymentMethod->default != 1)
+                                        <div class="d-grid gap-2">
+                                            <a class="btn btn-success" role="button" href="{{ route('app_update_payment_methods', [
+                                                'group' => 'payment_method',
+                                                'id' => $entreprise->id,
+                                                'id2' => $functionalUnit->id,
+                                                'id3' => $paymentMethod->id
+                                                ]) }}">
+                                                <i class="fa-solid fa-pen-to-square"></i>
+                                                {{ __('entreprise.edit') }}
+                                            </a>
+                                        </div>
+                                    @else
+                                        <div class="d-grid gap-2">
+                                            <button class="btn btn-success" type="button" disabled>
+                                                <i class="fa-solid fa-pen-to-square"></i>
+                                                {{ __('entreprise.edit') }}
+                                            </button>
+                                        </div>
+                                    @endif
+                                </div>
 
-                            <div class="col-md-6 mb-3">
-                                @if ($paymentMethod->default != 1)
-                                    <div class="d-grid gap-2">
-                                        <a class="btn btn-success" role="button" href="{{ route('app_update_payment_methods', [
-                                            'id' => $entreprise->id,
-                                            'id2' => $functionalUnit->id,
-                                            'id3' => $paymentMethod->id
-                                            ]) }}">
-                                            <i class="fa-solid fa-pen-to-square"></i>
-                                            {{ __('entreprise.edit') }}
-                                        </a>
-                                    </div>
-                                @else
-                                    <div class="d-grid gap-2">
-                                        <button class="btn btn-success" type="button" disabled>
-                                            <i class="fa-solid fa-pen-to-square"></i>
-                                            {{ __('entreprise.edit') }}
-                                        </button>
-                                    </div>
-                                @endif
-                            </div>
-
-                            <div class="col-md-6 mb-3">
-                                @if ($paymentMethod->default != 1 && !$encaissement_exit)
-                                    <div class="d-grid gap-2">
-                                        <button class="btn btn-danger" type="button" onclick="deleteElementThreeVal('{{ $paymentMethod->id }}', {{ $entreprise->id }}, {{ $functionalUnit->id }}, '{{ route('app_delete_payment_methods') }}', '{{ csrf_token() }}');" title="{{ __('entreprise.delete') }}">
-                                            <i class="fa-solid fa-trash-can"></i>
-                                            {{ __('entreprise.delete') }}
-                                        </button>
-                                    </div>
-                                @else
-                                    <div class="d-grid gap-2">
-                                        <button class="btn btn-danger" type="button" title="{{ __('entreprise.delete') }}" disabled>
-                                            <i class="fa-solid fa-trash-can"></i>
-                                            {{ __('entreprise.delete') }}
-                                        </button>
-                                    </div>
-                                @endif
-                            </div>
-
+                                <div class="col-md-6 mb-3">
+                                    @if ($paymentMethod->default != 1 && !$encaissement_exit)
+                                        <div class="d-grid gap-2">
+                                            <button class="btn btn-danger" type="button" onclick="deleteElementThreeVal('{{ $paymentMethod->id }}', {{ $entreprise->id }}, {{ $functionalUnit->id }}, '{{ route('app_delete_payment_methods') }}', '{{ csrf_token() }}');" title="{{ __('entreprise.delete') }}">
+                                                <i class="fa-solid fa-trash-can"></i>
+                                                {{ __('entreprise.delete') }}
+                                            </button>
+                                        </div>
+                                    @else
+                                        <div class="d-grid gap-2">
+                                            <button class="btn btn-danger" type="button" title="{{ __('entreprise.delete') }}" disabled>
+                                                <i class="fa-solid fa-trash-can"></i>
+                                                {{ __('entreprise.delete') }}
+                                            </button>
+                                        </div>
+                                    @endif
+                                </div>
+                            @endif
                         </div>
 
                     </div>
