+++ resources/views/service/info_category-service.blade.php
@@ -20,12 +20,12 @@
                 <div class="row">
                     <div class="col-12 col-md-6 order-md-1 order-last">
                         <h3>{{ __('service.service_category_details') }}</h3>
-                        <p class="text-subtitle text-muted"></p> 
+                        <p class="text-subtitle text-muted"></p>
                     </div>
                     <div class="col-12 col-md-6 order-md-2 order-first">
                         <nav class="float-start float-lg-end" style="--bs-breadcrumb-divider: url(&#34;data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' width='8' height='8'%3E%3Cpath d='M2.5 0L1 1.5 3.5 4 1 6.5 2.5 8l4-4-4-4z' fill='%236c757d'/%3E%3C/svg%3E&#34;);" aria-label="breadcrumb">
                             <ol class="breadcrumb">
-                              <li class="breadcrumb-item"><a href="{{ route('app_category_service', ['id' => $entreprise->id, 'id2' => $functionalUnit->id]) }}">{{ __('service.service_category') }}</a></li>
+                              <li class="breadcrumb-item"><a href="{{ route('app_category_service', ['group' => 'service', 'id' => $entreprise->id, 'id2' => $functionalUnit->id]) }}">{{ __('service.service_category') }}</a></li>
                               <li class="breadcrumb-item active" aria-current="page">{{ __('service.service_category_details') }}</li>
                             </ol>
                         </nav>
@@ -64,7 +64,7 @@
                                         ->where('id_cat', $category_service->id)
                                         ->count();
                         @endphp
-                       
+
                         <div class="row mb-4">
                             <div class="col-md-4">
                                 {{ __('dashboard.services') }}
@@ -93,7 +93,7 @@
                                 {{ __('client.creation_date') }}
                             </div>
                             <div class="col-md-8 text-primary fw-bold">
-                                {{ $category_service->created_at }} 
+                                {{ $category_service->created_at }}
                             </div>
                         </div>
 
@@ -102,33 +102,36 @@
                                 {{ __('client.last_modification_date') }}
                             </div>
                             <div class="col-md-8 text-primary fw-bold">
-                                {{ $category_service->updated_at }} 
+                                {{ $category_service->updated_at }}
                             </div>
                         </div>
 
                         <div class="row">
-
-                            <div class="col-md-6 mb-3">
-                                <div class="d-grid gap-2">
-                                    <a class="btn btn-success" role="button" href="{{ route('app_update_service_category', [
-                                        'id' => $entreprise->id, 
-                                        'id2' => $functionalUnit->id,
-                                        'id3' => $category_service->id
-                                        ]) }}">
-                                        <i class="fa-solid fa-pen-to-square"></i>
-                                        {{ __('entreprise.edit') }}
-                                    </a>
+                            @if ($permission_assign || Auth::user()->role->name == "admin" || Auth::user()->role->name == "superadmin")
+                                <div class="col-md-6 mb-3">
+                                    <div class="d-grid gap-2">
+                                        <a class="btn btn-success" role="button" href="{{ route('app_update_service_category', [
+                                            'group' => 'service',
+                                            'id' => $entreprise->id,
+                                            'id2' => $functionalUnit->id,
+                                            'id3' => $category_service->id
+                                            ]) }}">
+                                            <i class="fa-solid fa-pen-to-square"></i>
+                                            {{ __('entreprise.edit') }}
+                                        </a>
+                                    </div>
                                 </div>
-                            </div>
 
-                            <div class="col-md-6 mb-3">
-                                <div class="d-grid gap-2">
-                                    <button class="btn btn-danger" type="button" onclick="deleteElementThreeVal('{{ $category_service->id }}', {{ $entreprise->id }}, {{ $functionalUnit->id }}, '{{ route('app_delete_category_service') }}', '{{ csrf_token() }}');" title="{{ __('entreprise.delete') }}">
-                                        <i class="fa-solid fa-trash-can"></i>
-                                        {{ __('entreprise.delete') }}
-                                    </button>
+                                <div class="col-md-6 mb-3">
+                                    <div class="d-grid gap-2">
+                                        <button class="btn btn-danger" type="button" onclick="deleteElementThreeVal('{{ $category_service->id }}', {{ $entreprise->id }}, {{ $functionalUnit->id }}, '{{ route('app_delete_category_service') }}', '{{ csrf_token() }}');" title="{{ __('entreprise.delete') }}">
+                                            <i class="fa-solid fa-trash-can"></i>
+                                            {{ __('entreprise.delete') }}
+                                        </button>
+                                    </div>
                                 </div>
-                            </div>
+
+                            @endif
 
                         </div>
